wip: Start evaulating idea of templating w/ dhall
Config files are more or less fine, having dhall template a shell file
gets a little bit weird.

They both like to use `${…}` to interpolate, so would need to escape in
the dhall string literal to defer evaluation to runtime, not
template-time, which is a little gross, but there's probably a clever
way to circumvent this-- is that a good thing, though?